move/fix tests
Move the management tests to the input package and make all tests
pass.

This commit also makes the tests way faster to run and sends the
events to io.Discard instead of a file.
